+++ CHANGELOG.md
@@ -23,6 +23,7 @@ Change Log
 - Optimize syntax parsing.
 - Deprecate the feature to replace `$LN` in the outline menu template with the line number of the occurrence.
 - Prefer using .yml for syntax definition files over .yaml.
+- Add a temporal highlight for the current match in the editor (thanks to Ethan Wong!).
 - Remove original document icons for CoffeeScript and Tcl. 
 - [trivial] Accept script files for the Script menu with an uppercased file extension.
 - [trivial] Replace `\n` with `\R` for the new line meta character in the regular expression reference.